Added error call to check error handling (using bad SQL -> selct syntax).
Added redundant call request for connections. Blocked connection release in one of the threads.

Added checks for in_use_conn to not be negative.

Added check in release method to make sure that the number of available connections does not exceed maxconn. If the pool is full, the connection is not added back to the pool, but is closed.

Cleanup_pool method now closes redundant connections only if the number of available connections exceeds the minimum number of connections (minconn). This helps keep the pool at an optimal level.

Added check in release method to make sure that the number of available connections (all_connections) does not exceed maxconn. If the pool is full, the connection is not added back to the pool, but is closed.

Added that if a connection is closed by handle_connection_error, it will not be released again in finally.